+++ src/main/java/kr/hhplus/be/server/reservationInfo/domain/ConcertSchedule.java
@@ -1,12 +1,11 @@
-package kr.hhplus.be.server.reservation.domain;
+package kr.hhplus.be.server.reservationInfo.domain;
 
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import java.time.LocalDateTime;
 import kr.hhplus.be.server.common.persistence.BaseTimeEntity;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -19,22 +18,15 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Entity
-public class Reservation extends BaseTimeEntity {
+public class ConcertSchedule extends BaseTimeEntity {
 
     @Id
     private Long id;
 
-    private Long userId;
-
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "concert_schedule_id", nullable = false)
-    private ConcertSchedule concertSchedule;
+    @JoinColumn(name="concert_id", nullable=false)
+    private Concert concert;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "seat_id", nullable = false)
-    private Seat seat;
-
-    @Enumerated(EnumType.STRING)
-    private ReservationStatus status;
+    private LocalDateTime startAt;
 
 }
