+++ src/main/java/kr/hhplus/be/server/reservationInfo/domain/Reservation.java
@@ -1,12 +1,14 @@
-package kr.hhplus.be.server.reservation.domain;
-
+package kr.hhplus.be.server.reservationInfo.domain;
 
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import kr.hhplus.be.server.common.persistence.BaseTimeEntity;
+import kr.hhplus.be.server.reservation.domain.ReservationStatus;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -18,17 +20,22 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Entity
-public class Seat extends BaseTimeEntity {
+public class Reservation extends BaseTimeEntity {
 
     @Id
     private Long id;
 
+    private Long userId;
+
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name="concert_schedule_id", nullable=false)
+    @JoinColumn(name = "concert_schedule_id", nullable = false)
     private ConcertSchedule concertSchedule;
 
-    private Integer seatNo;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "seat_id", nullable = false)
+    private Seat seat;
 
-    private Long price;
+    @Enumerated(EnumType.STRING)
+    private ReservationStatus status;
 
 }
