+++ src/main/java/kr/hhplus/be/server/reservationInfo/dto/response/AvailableSeatsResponse.java
@@ -1,7 +1,7 @@
-package kr.hhplus.be.server.reservation.dto.response;
+package kr.hhplus.be.server.reservationInfo.dto.response;
 
 import java.util.List;
-import kr.hhplus.be.server.reservation.domain.Seat;
+import kr.hhplus.be.server.reservationInfo.domain.Seat;
 
 public record AvailableSeatsResponse(
     long scheduleId,
