+++ src/main/java/kr/hhplus/be/server/reservationInfo/repository/ConcertScheduleRepository.java
@@ -1,7 +1,7 @@
-package kr.hhplus.be.server.reservation.repository;
+package kr.hhplus.be.server.reservationInfo.repository;
 
 import java.util.List;
-import kr.hhplus.be.server.reservation.domain.ConcertSchedule;
+import kr.hhplus.be.server.reservationInfo.domain.ConcertSchedule;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ConcertScheduleRepository extends JpaRepository<ConcertSchedule, Long> {
