+++ src/main/java/kr/hhplus/be/server/reservationInfo/repository/SeatRepository.java
@@ -1,7 +1,7 @@
-package kr.hhplus.be.server.reservation.repository;
+package kr.hhplus.be.server.reservationInfo.repository;
 
 import java.util.List;
-import kr.hhplus.be.server.reservation.domain.Seat;
+import kr.hhplus.be.server.reservationInfo.domain.Seat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SeatRepository extends JpaRepository<Seat, Long> {
